rename: 트리 위치 검색 컴포넌트 폴더 구조 변경
