+++ src/components/TreeRegi/Location/Search/Result/style.ts
@@ -23,6 +23,7 @@ export const ResultItemAddress = styled.p`
 `;
 
 export const AddressType = styled.span`
+  color: var(--grey-dark);
   border-radius: 5px;
   padding: 1px 6px;
   background-color: var(--grey-light);
